Organize external overlay files
