+++ apps/web/src/routes/+layout.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import "../app.css";
-  import ExternalSiteOverlay from "$lib/ExternalSiteOverlay.svelte";
+  import ExternalSiteOverlay from "$lib/external-site-overlay/ExternalSiteOverlay.svelte";
 </script>
 
 <main class="min-h-screen">
